Totally untested (and unused) first draft for pure combat core rules. Just a bunch of static functions, that do or apply calculations from the rulebook. Next step would be to test them and integrate them into some kind of "combat flow" mechanism.

+++ ScarletGamma/src/core/Property.hpp
@@ -53,7 +53,7 @@ public:
 	/// \brief Interprets the value as formula/number and returns the result
 	///		after evaluation.
 	/// \throws Exception::NotEvaluateable
-	float Evaluate() const;
+	int Evaluate() const;
 
 	/// \brief Read access to the value string
 	const std::string& Value() const;
